+++ CHANGES.rst
@@ -11,6 +11,7 @@ Features:
 Improvements:
  - Add option to choose default media source (#2763)
  - Add option to choose default photo compression (#2763)
+ - Add option to disable camera shutter sound
 
 Other changes:
  -